requirements.txt
google-genai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
xhtml2pdf>=0.2.11
reportlab>=4.0.0
jinja2>=3.1.0
pyyaml>=6.0.1
pypdf>=4.0.0
click>=8.1.0
lxml>=5.0.0
